Switch data generation into one run
