--- data/data.csv
@@ -1,11 +1,0 @@
-title, host, city, photoURL
-input compress,Claudia Becker,Port Reid,https://s3.amazonaws.com/fec-overview-service-images/home_5.jpg
-Indian Rupee Ngultrum program,Haven Roob,Wuckertside,https://s3.amazonaws.com/fec-overview-service-images/home_5.jpg
-innovate Virgin Islands, British,Alejandrin Fay,South Derrickhaven,https://s3.amazonaws.com/fec-overview-service-images/home_5.jpg
-Auto Loan Account,Cielo Mills,Greenfelderberg,https://s3.amazonaws.com/fec-overview-service-images/home_5.jpg
-Tasty Borders Developer,Federico Pacocha,McKenziebury,https://s3.amazonaws.com/fec-overview-service-images/home_5.jpg
-Pennsylvania Creative,Edwardo Moore DVM,McDermottborough,https://s3.amazonaws.com/fec-overview-service-images/home_5.jpg
-methodical,Seth Keebler,Aronstad,https://s3.amazonaws.com/fec-overview-service-images/home_5.jpg
-European Unit of Account 17(E.U.A.-17) District,Carmen Murray,Rocioborough,https://s3.amazonaws.com/fec-overview-service-images/home_5.jpg
-Marketing,Jess Moore V,Lake Sandy,https://s3.amazonaws.com/fec-overview-service-images/home_5.jpg
-Central,Annalise Nicolas,New Gabriellashire,https://s3.amazonaws.com/fec-overview-service-images/home_5.jpg
